chore: move contracts into their own folder
