+++ src/ape/api/accounts.py
@@ -10,10 +10,10 @@ from ape.utils import cached_property
 
 from .address import AddressAPI
 from .base import abstractdataclass, abstractmethod
-from .contracts import ContractContainer, ContractInstance
 from .providers import ReceiptAPI, TransactionAPI, TransactionType
 
 if TYPE_CHECKING:
+    from ape.contracts import ContractContainer, ContractInstance
     from ape.managers.config import ConfigManager
 
 
@@ -136,7 +136,7 @@ class AccountAPI(AddressAPI):
 
         return self.call(txn, send_everything=value is None)
 
-    def deploy(self, contract: ContractContainer, *args, **kwargs) -> ContractInstance:
+    def deploy(self, contract: "ContractContainer", *args, **kwargs) -> "ContractInstance":
 
         txn = contract(*args, **kwargs)
         txn.sender = self.address
