+++ src/ape/contracts/__init__.py
@@ -1,0 +1,8 @@
+from .base import ContractContainer, ContractInstance, ContractLog, _Contract
+
+__all__ = [
+    "_Contract",
+    "ContractContainer",
+    "ContractInstance",
+    "ContractLog",
+]
