+++ src/ape/contracts/base.py
@@ -2,13 +2,12 @@ from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 from eth_utils import to_bytes
 
+from ape.api.address import Address, AddressAPI
+from ape.api.base import dataclass
+from ape.api.providers import ProviderAPI, ReceiptAPI, TransactionAPI
 from ape.exceptions import ArgumentsLengthError, ContractDeployError, TransactionError
 from ape.logging import logger
 from ape.types import ABI, AddressType, ContractType
-
-from .address import Address, AddressAPI
-from .base import dataclass
-from .providers import ProviderAPI, ReceiptAPI, TransactionAPI
 
 if TYPE_CHECKING:
     from ape.managers.converters import ConversionManager
